Add timer in RoleViewModel

+++ SpyFall/app/src/main/java/com/example/spyfall/data/entity/Game.kt
@@ -3,6 +3,6 @@ package com.example.spyfall.data.entity
 data class Game(
     val host: String? = null,
     val status: GameStatus? = null,
-    val duration: Int? = null,
+    val duration: Long? = null,
     val vote: String? = null
 )
